Issue#32 fixed profile names
