+++ src/FluentSerializer.Core.BenchmarkUtils/Profiles/WriteProfile.cs
@@ -1,14 +1,14 @@
-using BenchmarkDotNet.Attributes;
-using FluentSerializer.Core.DataNodes;
-using FluentSerializer.Core.TestUtils.Helpers;
 using System;
 using System.IO;
 using System.Text;
+using BenchmarkDotNet.Attributes;
+using FluentSerializer.Core.DataNodes;
+using FluentSerializer.Core.TestUtils.Helpers;
 
-namespace FluentSerializer.Core.BenchmarkUtils.Profilers
+namespace FluentSerializer.Core.BenchmarkUtils.Profiles
 {
 	[MemoryDiagnoser]
-    public abstract class ReadProfile
+    public abstract class WriteProfile
     {
         private MemoryStream? _writeStream;
         private StreamWriter? _streamWriter;
@@ -37,9 +37,10 @@ namespace FluentSerializer.Core.BenchmarkUtils.Profilers
             _writeStream = null;
         }
 
-        public virtual void Read(IDataNode value)
+        public void Write(IDataNode value)
         {
             value.WriteTo(TestStringBuilderPool.StringBuilderPool, _streamWriter!, true);
+
             _streamWriter!.Flush();
             _ = Encoding.UTF8.GetString(_writeStream!.ToArray());
         }
