+++ src/FluentSerializer.Json.Benchmark/Profiles/JsonWriteProfile.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
-using System.IO;
 using BenchmarkDotNet.Attributes;
-using FluentSerializer.Core.BenchmarkUtils.Profilers;
+using FluentSerializer.Core.DataNodes;
+using FluentSerializer.Core.BenchmarkUtils.Profiles;
 using FluentSerializer.Core.BenchmarkUtils.TestData;
 using FluentSerializer.Json.Benchmark.Data;
 
@@ -9,12 +9,12 @@ namespace FluentSerializer.Json.Benchmark.Profiles
 {
     public class JsonWriteProfile : WriteProfile
     {
-        public IEnumerable<TestCase<Stream>> Values => JsonDataCollection.Default.StringTestData;
+        public IEnumerable<TestCase<IDataNode>> Values() => JsonDataCollection.Default.ObjectTestData;
 
         [ParamsSource(nameof(Values))]
-        public TestCase<Stream> Value { get => CaseValue; set => CaseValue = value; }
+        public TestCase<IDataNode> Value { get; set; }
 
-        [Benchmark, BenchmarkCategory(nameof(Write))]
-        public void Write() => JsonParser.Parse(CaseReader.ReadToEnd());
+        [Benchmark, BenchmarkCategory(nameof(WriteJson))]
+        public void WriteJson() => Write(Value.GetData());
     }
 }
