+++ src/FluentSerializer.Xml.Benchmark/Profiles/XmlWriteProfile.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
-using System.IO;
 using BenchmarkDotNet.Attributes;
-using FluentSerializer.Core.BenchmarkUtils.Profilers;
+using FluentSerializer.Core.DataNodes;
+using FluentSerializer.Core.BenchmarkUtils.Profiles;
 using FluentSerializer.Core.BenchmarkUtils.TestData;
 using FluentSerializer.Xml.Benchmark.Data;
 
@@ -9,12 +9,12 @@ namespace FluentSerializer.Xml.Benchmark.Profiles
 {
     public class XmlWriteProfile : WriteProfile
     {
-        public IEnumerable<TestCase<Stream>> Values => XmlDataCollection.Default.StringTestData;
+        public IEnumerable<TestCase<IDataNode>> Values() => XmlDataCollection.Default.ObjectTestData;
 
         [ParamsSource(nameof(Values))]
-        public TestCase<Stream> Value { get => CaseValue; set => CaseValue = value; }
+        public TestCase<IDataNode> Value { get; set; }
 
         [Benchmark, BenchmarkCategory(nameof(Write))]
-        public void Write() => XmlParser.Parse(CaseReader.ReadToEnd());
+        public void WriteXml() => Write(Value.GetData());
     }
 }
